packages/unifi-mcp-relay/pyproject.toml
[project]
name = "unifi-mcp-relay"
dynamic = ["version"]
description = "MCP relay: bridges local MCP servers to a Cloudflare Worker gateway"
requires-python = ">=3.13"
dependencies = [
    "websockets>=13.0",
    "mcp[cli]>=1.26.0,<2",
    "aiohttp>=3.8.5",
    "python-dotenv>=1.0.0",
]

[project.scripts]
unifi-mcp-relay = "unifi_mcp_relay.__main__:main"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "aioresponses>=0.7.0",
]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "vcs"
raw-options.root = "../.."
raw-options.tag_regex = "^relay/v(?P<version>\\d+(?:\\.\\d+)*)(?:\\S*)$"
raw-options.git_describe_command = ["git", "describe", "--dirty", "--tags", "--long", "--match", "relay/v*"]
raw-options.fallback_version = "0.0.0"

[tool.hatch.build.hooks.vcs]
version-file = "src/unifi_mcp_relay/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/unifi_mcp_relay"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
